Don't delete old game cache if it's in use by another server

+++ src-tauri/src/lib.rs
@@ -429,17 +429,17 @@ async fn prep_launch(
                                 "Upgraded cache from parent version {} for {}",
                                 parent_uuid, version_uuid
                             );
-                            if state.config.launcher.delete_old_game_caches {
+
+                            if state.config.launcher.delete_old_game_caches
+                                && state.get_version_use_count(parent_uuid) == 0
+                            {
                                 if let Err(e) = util::delete_dir(&parent_cache_dir) {
                                     warn!(
                                         "Failed to delete cache for parent version {}: {}",
                                         parent_uuid, e
                                     );
                                 } else {
-                                    info!(
-                                        "Deleted cache for parent version {}",
-                                        parent_uuid
-                                    );
+                                    info!("Deleted cache for parent version {}", parent_uuid);
                                 }
                             }
                         }
